requirements.txt
python-telegram-bot
openai
duckduckgo-search
lancedb
numpy
pandas
google-generativeai
phidata
pypdf